README: clarify, add video, add common issues, refactor
